fix(dev): move generated prisma clients outside watched src
