+++ services/budget-api/src/lib/prisma.ts
@@ -1,3 +1,3 @@
-import { PrismaClient } from '../generated/prisma/index.js';
+import { PrismaClient } from '../../generated/prisma/index.js';
 
 export const prisma = new PrismaClient();
